Got rid of use of bytes as values may need to go -ve when cycling

+++ Shuffle.h
@@ -20,7 +20,7 @@ public:
   void reset();
 private:
   unsigned long gate;
-  byte beat;
+  int beat;
   ShuffleState state;
   Signal shuffleClock(int track, int shuffle);
   void newCycle();
